requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
psycopg2==2.9.9
requests==2.31.0
SQLAlchemy==2.0.23
geomet==1.1.0
qwc-services-core==1.3.24
